adding beamdyn file writing to akshat's solid model example.
